+++ Perceptron.h
@@ -6,9 +6,9 @@
 
 double* initTraining(int, int);
 double generateDouble();
-double* startTraining(double*, double*, int*, int, int);
-void training(double*, double*, int*, int, int);
-int computeOutput(double*, double*, int, int);
+double* startTraining(double*, double*, double*, int, int);
+void training(double*, double*, double*, int, int);
+double computeOutput(double*, double*, int, int);
 int output(double*, double*, int);
 void updateWeights(double*, double*, double, int, int);
 void freeWeights(double*);
